chore: set and use AuthContext for authentication states and logic
