+++ frontend/src/App.jsx
@@ -1,66 +1,17 @@
-import { useState, useEffect } from 'react';
 import { Routes, Route, Navigate } from 'react-router-dom';
+import { useContext } from 'react'
+import { AuthContext } from './context/AuthContext';
 import Header from './components/Header';
 import TaskList from './components/TaskList';
 import Login from './components/Login';
 import Register from './components/Register';
-import { getApiUrl } from './utils/api';
 
 const App = () => {
-  // 1- state variables
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [isLoading, setIsLoading] = useState(true);   // Add loading state
+  // State variables
+  const { isAuthenticated, isLoading } = useContext(AuthContext);
 
 
-  // 2- useEffect logic definition
-  useEffect(() => {
-    const verifyToken = async () => {
-
-      // await new Promise(resolve => setTimeout(resolve, 2000));  // to test the spinner only
-
-      const token = localStorage.getItem('token');
-      if (token) {
-        try {
-          const response = await fetch(getApiUrl('/api/auth/verify'), {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-          });
-
-          if (response.ok) {
-            setIsAuthenticated(true);
-          } else {
-            localStorage.removeItem('token');
-            setIsAuthenticated(false);
-          }
-
-        } catch (error) {
-          localStorage.removeItem('token');
-          setIsAuthenticated(false);
-        }
-      } else {
-        setIsAuthenticated(false);
-      }
-      setIsLoading(false); // Set loading to false when verification is complete
-    };
-
-    verifyToken();
-  }, []);
-
-
-  // 3- handle functions definitions
-  const handleLogin = () => {
-    setIsAuthenticated(true);
-  };
-
-
-  const handleLogout = () => {
-    localStorage.removeItem('token');
-    setIsAuthenticated(false);
-  };
-
-
-  // 4- Don't render anything while verifying token (inside useEffect)
+  // Don't render anything while verifying token (inside useEffect)
   const LoadingSpinner = () => (
     <div className="flex justify-center items-center h-screen">
       <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-blue-500"></div>
@@ -72,13 +23,10 @@ const App = () => {
   }
 
 
-  // 5- Render Main app only after isLoading is false
+  // Render Main app only after isLoading is false
   return (
     <div className="max-w-4xl mx-auto p-4 sm:max-w-2xl md:max-w-4xl lg:max-w-4xl">
-      <Header
-        isAuthenticated={isAuthenticated}
-        onLogout={handleLogout}
-      />
+      <Header />
       <Routes>
         <Route
           path="/"
@@ -86,7 +34,7 @@ const App = () => {
         />
         <Route
           path="/login"
-          element={<Login onLogin={handleLogin} />}
+          element={<Login />}
         />
         <Route
           path="/register"
