+++ frontend/src/main.jsx
@@ -1,13 +1,17 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { AuthProvider } from './context/AuthContext.jsx'
+import { BrowserRouter as Router} from 'react-router-dom'
 import './index.css'
 import App from './App.jsx'
-import { BrowserRouter as Router} from 'react-router-dom'
+
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <Router>
-      <App />
-    </Router>
+    <AuthProvider>
+      <Router>
+        <App />
+      </Router>
+    </AuthProvider>
   </StrictMode>,
 )
